==> 1_SuppliersDupsAnalysis.ipynb <==
### Step 1. Load required packages

import pandas as pd
import pandasql as ps

### Step 2 - Loads suppliers interface data


from custom_procedures import newestFile
sourceFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/1_Suppliers_Interface'
workFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/Work/1_Suppliers_Interface'
sourceFile = newestFile(sourceFolder)
if sourceFile:
    print(sourceFile)
    df = pd.read_csv(sourceFile, sep=';', quotechar=chr(39), encoding='latin-1')
    print('=> Columns loaded')
    for columnName in df.columns: 
        print(columnName)
print(len(df.index),' records loaded')

### Step 3 - Find duplicates by NUM_1099

sql_query = '''
SELECT NUM_1099,numObs FROM
(SELECT NUM_1099,COUNT(*) AS numObs FROM df
GROUP BY NUM_1099)
ORDER BY numObs DESC
'''
dupsAnalysisDf = ps.sqldf(sql_query)
print('Total number of records, ', len(dupsAnalysisDf.index))
#
sql_query = '''
SELECT * FROM dupsAnalysisDf
WHERE numObs=1
ORDER BY NUM_1099 ASC
'''
uniqueRecordsDf = ps.sqldf(sql_query)
print('Total number unique records, ', len(uniqueRecordsDf.index))
#
sql_query = '''
SELECT * FROM dupsAnalysisDf
WHERE numObs>1
ORDER BY numObs DESC
'''
duplicateRecordsDf = ps.sqldf(sql_query)
print('Total number of duplicated unique records, ', len(duplicateRecordsDf.index))

### Step 4 - Asigns a case number to each duplicate record

duplicateRecordsDf['caseNum'] = 0
duplicateRecordsDf['caseNum'].astype(int)
print(duplicateRecordsDf.columns)
caseNum = 1
for index, row in duplicateRecordsDf.iterrows():
    duplicateRecordsDf.loc[index, 'caseNum'] = caseNum
    caseNum +=1
print(caseNum, ' Number of cases to be inspected.')
print(duplicateRecordsDf)

### Step 5 - Validate the number of duplicater per cases matches the total number of records

numRecordsVal = 0
for index, row in duplicateRecordsDf.iterrows():
    numRecordsVal = numRecordsVal + int(row['numObs'])
print(numRecordsVal, ' records contained in ', caseNum, ' cases.')

### Step 6 - Save dataframes to csv files

from pathlib import Path
uniqueRecordsDf.to_csv(str(Path(workFolder,'uniqueRecordsDf.csv', encoding='utf-8')))
duplicateRecordsDf.to_csv(str(Path(workFolder,'duplicateRecordsDf.csv', encoding='utf-8')))

### Step 7 - Create a reduced dataframe to identify cases

from pathlib import Path
reducedDf = df[['SISTEMA_LEGADO','NUMERO_PROVEEDOR','VENDOR_NAME','NUM_1099']]
reducedDf.to_csv(str(Path(workFolder,'reducedDf.csv', encoding='utf-8')))
print(reducedDf)

### Step 8 - Add auxiliary columns for case identification

duplicateRecordsDf['multiple_feeds'] = False
duplicateRecordsDf['multiple_feeds'].astype(bool)
duplicateRecordsDf['multiples_ids'] = False
duplicateRecordsDf['multiples_ids'].astype(bool)
duplicateRecordsDf['multiples_names'] = False
duplicateRecordsDf['multiples_names'].astype(bool)

### Step 9 - Validating conditions

from pathlib import Path
for index, row in duplicateRecordsDf.iterrows():
    subset = reducedDf[reducedDf['NUM_1099'] == row['NUM_1099']]
    # 'NUMERO_PROVEEDOR','VENDOR_NAME','NUM_1099'
    num_feeds = len(subset.SISTEMA_LEGADO.value_counts())
    num_ids = len(subset.NUMERO_PROVEEDOR.value_counts())
    num_names = len(subset.VENDOR_NAME.value_counts())
    if num_feeds > 1:
        duplicateRecordsDf.loc[index, 'multiple_feeds'] = True
    if num_ids > 1:
        duplicateRecordsDf.loc[index, 'multiples_ids'] = True
    if num_names > 1:
        duplicateRecordsDf.loc[index, 'multiples_names'] = True
    #if index >= 100:
    #    break
#print(duplicateRecordsDf.head(100))
reducedDf.to_csv(str(Path(workFolder,'duplicateRecordsWithClassificationDf.csv', encoding='utf-8')))
