==> 1b_SuppliersDQ&Scenarios.ipynb <==
### Step 1. Load required packages

import pandas as pd
import pandasql as ps
from pathlib import Path
workFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/Work/1_Suppliers_Interface'

### Step 2. Read required data

reducedDf = pd.read_csv(Path(workFolder,'reducedDf.csv'), encoding='utf-8')
duplicateRecordsWithClassificationDf = pd.read_csv(Path(workFolder,'duplicateRecordsWithClassificationDf.csv'), encoding='utf-8')

### Step 3 - Identify Scenarios - Cases with more observations

sql_query = '''
SELECT numObs,COUNT(*) AS numCases FROM duplicateRecordsWithClassificationDf
GROUP BY numObs
ORDER BY numCases DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['numObs'],row['numCases'])

### Step 4 - Analisis of specific scenario

sql_query = '''
SELECT multiple_feeds,multiples_ids,multiples_names,COUNT(*) AS numCases FROM
(SELECT * FROM duplicateRecordsWithClassificationDf
WHERE numObs=2)
GROUP BY multiple_feeds,multiples_ids,multiples_names
ORDER BY numCases DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['multiple_feeds'], row['multiples_ids'], row['multiples_names'], row['numCases'])

print(reducedDf.head(100))

sql_query = '''
SELECT B.SISTEMA_LEGADO,B.NUMERO_PROVEEDOR,B.VENDOR_NAME,A.NUM_1099,
A.numObs,A.caseNum,A.multiple_feeds,A.multiples_ids,A.multiples_names
FROM
(SELECT NUM_1099,numObs,caseNum,multiple_feeds,multiples_ids,multiples_names
FROM duplicateRecordsWithClassificationDf
WHERE numObs=2 AND multiple_feeds=1 AND multiples_ids=0 AND multiples_names=0) AS A
LEFT JOIN
(SELECT SISTEMA_LEGADO,NUMERO_PROVEEDOR,VENDOR_NAME,NUM_1099 FROM reducedDf) AS B
ON (A.NUM_1099=B.NUM_1099)
ORDER BY A.NUM_1099
'''
result = ps.sqldf(sql_query)
print(result.head(10))
result.to_csv(str(Path(workFolder,'duplicateRecords_2Obs_multipleFeeds.csv', encoding='utf-8')))
