==> 3_SuppliersSites.ipynb <==
## Data from File 3 Suppliers Interface

### Step 1. Load required package

import pandas as pd
import pandasql as ps
from pathlib import Path

### Step 2 - Get sample from source file

sourceFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/3_Suppliers_Sites'
workFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/Work/5_Bank_Accounts'
fileName = '3. SUPPLIER_SITES_INTERFACE_COPPEL.csv'
inputTextFile = open(Path(sourceFolder, fileName), 'r', encoding='latin-1')
if inputTextFile:
    numTextLines = 0
    while True:
        numTextLines += 1
        textLine = inputTextFile.readline()
        print(textLine)
        if numTextLines >10:
            break
    inputTextFile.close()

### Step 3 - Load supplier paye

df = pd.read_csv(Path(sourceFolder,fileName), sep=';', quotechar=chr(39), encoding='latin-1')
print(df.head(10))
print(len(df.index), ' records found. ')

pd.set_option('display.max_rows', 250)
df.dtypes

### Step 4 - Creates a modified dataset with key field as string

modDf = df
modDf['SISTEMA_LEGADO']=modDf['SISTEMA_LEGADO'].astype(pd.StringDtype())
modDf['NUMERO_PROVEEDOR']=modDf['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
modDf['IMPORT_ACTION']=modDf['IMPORT_ACTION'].astype(pd.StringDtype())
modDf['VENDOR_NAME']=modDf['VENDOR_NAME'].astype(pd.StringDtype())
modDf['PROCUREMENT_BUSINESS_UNIT_NAME']=modDf['PROCUREMENT_BUSINESS_UNIT_NAME'].astype(pd.StringDtype())
modDf['PARTY_SITE_NAME']=modDf['PARTY_SITE_NAME'].astype(pd.StringDtype())
modDf['VENDOR_SITE_CODE']=modDf['VENDOR_SITE_CODE'].astype(pd.StringDtype())
modDf['RFQ_ONLY_SITE_FLAG']=modDf['RFQ_ONLY_SITE_FLAG'].astype(pd.StringDtype())
modDf['PURCHASING_SITE_FLAG']=modDf['PURCHASING_SITE_FLAG'].astype(pd.StringDtype())
modDf['PAY_SITE_FLAG']=modDf['PAY_SITE_FLAG'].astype(pd.StringDtype())
modDf['PRIMARY_PAY_SITE_FLAG']=modDf['PRIMARY_PAY_SITE_FLAG'].astype(pd.StringDtype())
print(modDf.head(10))
print(modDf.dtypes)

### Step 5 - Reads Data in Scope

dataInScopeDf = pd.read_csv(Path(workFolder, 'dataInScopeDfs.csv'), encoding='utf-8')
print(dataInScopeDf.head(10))

### Step 6 - Get Tax Id's or Supplier Id's in scope

sql_query = '''
SELECT NUMERO_PROVEEDOR FROM
(SELECT DISTINCT NUMERO_PROVEEDOR FROM dataInScopeDf)
ORDER BY NUMERO_PROVEEDOR ASC
'''
SuppliersInScope = ps.sqldf(sql_query)
SuppliersInScope['NUMERO_PROVEEDOR'] = SuppliersInScope['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
print(SuppliersInScope.dtypes)
print(len(SuppliersInScope.index), ' suppliers in scope.')
print(SuppliersInScope.head(10))

### Step 7 - Merge Suppliers in Scope with 5.1 Table data

subSetDf = SuppliersInScope.merge(modDf, on='NUMERO_PROVEEDOR', how='left')
subSetDf = subSetDf[subSetDf['VENDOR_SITE_CODE'].notna()]
print(len(subSetDf.index), ' records obtained.')
print(subSetDf.head(10))

### Step 8 - Validate Legacy Systems in Scope

legacySystemsFound = subSetDf.SISTEMA_LEGADO.unique()
print(len(legacySystemsFound), ' records found.')
print(list(legacySystemsFound))

### Step 9 - check distribution of records per legacy system

sql_query = '''
SELECT SISTEMA_LEGADO,COUNT(*) AS numObs FROM subSetDf
GROUP BY SISTEMA_LEGADO
ORDER BY numObs DESC
'''
distPerLegacySystemDf = ps.sqldf(sql_query)
for index, row in distPerLegacySystemDf.iterrows():
    print(row['SISTEMA_LEGADO'], ',', row['numObs'])

sql_query = '''
SELECT DISTINCT NUMERO_PROVEEDOR,VENDOR_SITE_CODE FROM subSetDf
'''
SupplierBUAndVSTDf = ps.sqldf(sql_query)
print(len(SupplierBUAndVSTDf.index), ' records in scope')
SupplierBUAndVSTDf.to_csv(Path(workFolder, 'SupplierBUAndVSTDf.csv'), encoding='utf-8')
sql_query = '''
SELECT NUMERO_PROVEEDOR,COUNT(VENDOR_SITE_CODE) AS numVSCs FROM SupplierBUAndVSTDf
GROUP BY NUMERO_PROVEEDOR
ORDER BY numVSCs DESC
'''
result = ps.sqldf(sql_query)
print(result.head(10))
sql_query = '''
SELECT numVSCs,COUNT(*) AS numObs FROM result
GROUP BY numVSCs
ORDER BY numObs DESC
'''
result2 = ps.sqldf(sql_query)
print(result2.head(10))

### Step 10 - Create new report structure

newReportDf = subSetDf[['SISTEMA_LEGADO','NUMERO_PROVEEDOR','BATCH_ID','IMPORT_ACTION','VENDOR_NAME','PROCUREMENT_BUSINESS_UNIT_NAME','PARTY_SITE_NAME','VENDOR_SITE_CODE','VENDOR_SITE_CODE_NEW','INACTIVE_DATE','RFQ_ONLY_SITE_FLAG','PURCHASING_SITE_FLAG','PCARD_SITE_FLAG','PAY_SITE_FLAG','PRIMARY_PAY_SITE_FLAG','TAX_REPORTING_SITE_FLAG','VENDOR_SITE_CODE_ALT','CUSTOMER_NUM','B2B_COMMUNICATION_METHOD','B2B_SITE_CODE','SUPPLIER_NOTIF_METHOD','EMAIL_ADDRESS','FAX_COUNTRY_CODE','FAX_AREA_CODE','FAX','HOLD_FLAG','PURCHASING_HOLD_REASON','CARRIER','MODE_OF_TRANSPORT_CODE','SERVICE_LEVEL_CODE','FREIGHT_TERMS_LOOKUP_CODE','PAY_ON_CODE','FOB_LOOKUP_CODE','COUNTRY_OF_ORIGIN_CODE','BUYER_MANAGED_TRANSPORT_FLAG','PAY_ON_USE_FLAG','AGING_ONSET_POINT','AGING_PERIOD_DAYS','CONSUMPTION_ADVICE_FREQUENCY','CONSUMPTION_ADVICE_SUMMARY','DEFAULT_PAY_SITE_CODE','PAY_ON_RECEIPT_SUMMARY_CODE','GAPLESS_INV_NUM_FLAG','SELLING_COMPANY_IDENTIFIER','CREATE_DEBIT_MEMO_FLAG','ENFORCE_SHIP_TO_LOCATION_CODE','RECEIVING_ROUTING_ID','QTY_RCV_TOLERANCE','QTY_RCV_EXCEPTION_CODE','DAYS_EARLY_RECEIPT_ALLOWED','DAYS_LATE_RECEIPT_ALLOWED','ALLOW_SUBSTITUTE_RECEIPTS_FLAG','ALLOW_UNORDERED_RECEIPTS_FLAG','RECEIPT_DAYS_EXCEPTION_CODE','INVOICE_CURRENCY_CODE','INVOICE_AMOUNT_LIMIT','MATCH_OPTION','MATCH_APPROVAL_LEVEL','PAYMENT_CURRENCY_CODE','PAYMENT_PRIORITY','PAY_GROUP_LOOKUP_CODE','TOLERANCE_NAME','SERVICES_TOLERANCE','HOLD_ALL_PAYMENTS_FLAG','HOLD_UNMATCHED_INVOICES_FLAG','HOLD_FUTURE_PAYMENTS_FLAG','HOLD_BY','PAYMENT_HOLD_DATE','HOLD_REASON','TERMS_NAME','TERMS_DATE_BASIS','PAY_DATE_BASIS_LOOKUP_CODE','BANK_CHARGE_DEDUCTION_TYPE','ALWAYS_TAKE_DISC_FLAG','EXCLUDE_FREIGHT_FROM_DISCOUNT','EXCLUDE_TAX_FROM_DISCOUNT','AUTO_CALCULATE_INTEREST_FLAG','VAT_CODE','VAT_REGISTRATION_NUM','PAYMENT_METHOD_LOOKUP_CODE','DELIVERY_CHANNEL_CODE','BANK_INSTRUCTION1_CODE','BANK_INSTRUCTION2_CODE','BANK_INSTRUCTION_DETAILS','SETTLEMENT_PRIORITY','PAYMENT_TEXT_MESSAGE1','PAYMENT_TEXT_MESSAGE2','PAYMENT_TEXT_MESSAGE3','IBY_BANK_CHARGE_BEARER','PAYMENT_REASON_CODE','PAYMENT_REASON_COMMENTS','REMIT_ADVICE_DELIVERY_METHOD','REMITTANCE_EMAIL','REMITTANCE_FAX','ATTRIBUTE_CATEGORY','ATTRIBUTE1','ATTRIBUTE2','ATTRIBUTE3','ATTRIBUTE4','ATTRIBUTE5','ATTRIBUTE6','ATTRIBUTE7','ATTRIBUTE8','ATTRIBUTE9','ATTRIBUTE10','ATTRIBUTE11','ATTRIBUTE12','ATTRIBUTE13','ATTRIBUTE14','ATTRIBUTE15','ATTRIBUTE16','ATTRIBUTE17','ATTRIBUTE18','ATTRIBUTE19','ATTRIBUTE20','ATTRIBUTE_DATE1','ATTRIBUTE_DATE2','ATTRIBUTE_DATE3','ATTRIBUTE_DATE4','ATTRIBUTE_DATE5','ATTRIBUTE_DATE6','ATTRIBUTE_DATE7','ATTRIBUTE_DATE8','ATTRIBUTE_DATE9','ATTRIBUTE_DATE10','ATTRIBUTE_TIMESTAMP1','ATTRIBUTE_TIMESTAMP2','ATTRIBUTE_TIMESTAMP3','ATTRIBUTE_TIMESTAMP4','ATTRIBUTE_TIMESTAMP5','ATTRIBUTE_TIMESTAMP6','ATTRIBUTE_TIMESTAMP7','ATTRIBUTE_TIMESTAMP8','ATTRIBUTE_TIMESTAMP9','ATTRIBUTE_TIMESTAMP10','ATTRIBUTE_NUMBER1','ATTRIBUTE_NUMBER2','ATTRIBUTE_NUMBER3','ATTRIBUTE_NUMBER4','ATTRIBUTE_NUMBER5','ATTRIBUTE_NUMBER6','ATTRIBUTE_NUMBER7','ATTRIBUTE_NUMBER8','ATTRIBUTE_NUMBER9','ATTRIBUTE_NUMBER10','GLOBAL_ATTRIBUTE_CATEGORY','GLOBAL_ATTRIBUTE1','GLOBAL_ATTRIBUTE2','GLOBAL_ATTRIBUTE3','GLOBAL_ATTRIBUTE4','GLOBAL_ATTRIBUTE5','GLOBAL_ATTRIBUTE6','GLOBAL_ATTRIBUTE7','GLOBAL_ATTRIBUTE8','GLOBAL_ATTRIBUTE9','GLOBAL_ATTRIBUTE10','GLOBAL_ATTRIBUTE11','GLOBAL_ATTRIBUTE12','GLOBAL_ATTRIBUTE13','GLOBAL_ATTRIBUTE14','GLOBAL_ATTRIBUTE15','GLOBAL_ATTRIBUTE16','GLOBAL_ATTRIBUTE17','GLOBAL_ATTRIBUTE18','GLOBAL_ATTRIBUTE19','GLOBAL_ATTRIBUTE20','GLOBAL_ATTRIBUTE_DATE1','GLOBAL_ATTRIBUTE_DATE2','GLOBAL_ATTRIBUTE_DATE3','GLOBAL_ATTRIBUTE_DATE4','GLOBAL_ATTRIBUTE_DATE5','GLOBAL_ATTRIBUTE_DATE6','GLOBAL_ATTRIBUTE_DATE7','GLOBAL_ATTRIBUTE_DATE8','GLOBAL_ATTRIBUTE_DATE9','GLOBAL_ATTRIBUTE_DATE10','GLOBAL_ATTRIBUTE_DATETIME1','GLOBAL_ATTRIBUTE_DATETIME2','GLOBAL_ATTRIBUTE_DATETIME3','GLOBAL_ATTRIBUTE_DATETIME4','GLOBAL_ATTRIBUTE_DATETIME5','GLOBAL_ATTRIBUTE_DATETIME6','GLOBAL_ATTRIBUTE_DATETIME7','GLOBAL_ATTRIBUTE_DATETIME8','GLOBAL_ATTRIBUTE_DATETIME9','GLOBAL_ATTRIBUTE_DATETIME10','GLOBAL_ATTRIBUTE_NUMBER1','GLOBAL_ATTRIBUTE_NUMBER2','GLOBAL_ATTRIBUTE_NUMBER3','GLOBAL_ATTRIBUTE_NUMBER4','GLOBAL_ATTRIBUTE_NUMBER5','GLOBAL_ATTRIBUTE_NUMBER6','GLOBAL_ATTRIBUTE_NUMBER7','GLOBAL_ATTRIBUTE_NUMBER8','GLOBAL_ATTRIBUTE_NUMBER9','GLOBAL_ATTRIBUTE_NUMBER10','PO_ACK_REQD_CODE','PO_ACK_REQD_DAYS','INVOICE_CHANNEL','PAYEE_SERVICE_LEVEL_CODE','EXCLUSIVE_PAYMENT_FLAG']]
print(len(newReportDf.index), ' in new report structure')




print(newReportDf.dtypes)

newReportDf.fillna('', inplace=True)
with open(Path(workFolder,'3. SUPPLIER_SITES_INTERFACE_COPPEL(NEW).csv'), 'w', encoding='utf-8') as textFile:
    tmpVector = ['SISTEMA_LEGADO','NUMERO_PROVEEDOR','BATCH_ID','IMPORT_ACTION','VENDOR_NAME','PROCUREMENT_BUSINESS_UNIT_NAME','PARTY_SITE_NAME','VENDOR_SITE_CODE','VENDOR_SITE_CODE_NEW','INACTIVE_DATE','RFQ_ONLY_SITE_FLAG','PURCHASING_SITE_FLAG','PCARD_SITE_FLAG','PAY_SITE_FLAG','PRIMARY_PAY_SITE_FLAG','TAX_REPORTING_SITE_FLAG','VENDOR_SITE_CODE_ALT','CUSTOMER_NUM','B2B_COMMUNICATION_METHOD','B2B_SITE_CODE','SUPPLIER_NOTIF_METHOD','EMAIL_ADDRESS','FAX_COUNTRY_CODE','FAX_AREA_CODE','FAX','HOLD_FLAG','PURCHASING_HOLD_REASON','CARRIER','MODE_OF_TRANSPORT_CODE','SERVICE_LEVEL_CODE','FREIGHT_TERMS_LOOKUP_CODE','PAY_ON_CODE','FOB_LOOKUP_CODE','COUNTRY_OF_ORIGIN_CODE','BUYER_MANAGED_TRANSPORT_FLAG','PAY_ON_USE_FLAG','AGING_ONSET_POINT','AGING_PERIOD_DAYS','CONSUMPTION_ADVICE_FREQUENCY','CONSUMPTION_ADVICE_SUMMARY','DEFAULT_PAY_SITE_CODE','PAY_ON_RECEIPT_SUMMARY_CODE','GAPLESS_INV_NUM_FLAG','SELLING_COMPANY_IDENTIFIER','CREATE_DEBIT_MEMO_FLAG','ENFORCE_SHIP_TO_LOCATION_CODE','RECEIVING_ROUTING_ID','QTY_RCV_TOLERANCE','QTY_RCV_EXCEPTION_CODE','DAYS_EARLY_RECEIPT_ALLOWED','DAYS_LATE_RECEIPT_ALLOWED','ALLOW_SUBSTITUTE_RECEIPTS_FLAG','ALLOW_UNORDERED_RECEIPTS_FLAG','RECEIPT_DAYS_EXCEPTION_CODE','INVOICE_CURRENCY_CODE','INVOICE_AMOUNT_LIMIT','MATCH_OPTION','MATCH_APPROVAL_LEVEL','PAYMENT_CURRENCY_CODE','PAYMENT_PRIORITY','PAY_GROUP_LOOKUP_CODE','TOLERANCE_NAME','SERVICES_TOLERANCE','HOLD_ALL_PAYMENTS_FLAG','HOLD_UNMATCHED_INVOICES_FLAG','HOLD_FUTURE_PAYMENTS_FLAG','HOLD_BY','PAYMENT_HOLD_DATE','HOLD_REASON','TERMS_NAME','TERMS_DATE_BASIS','PAY_DATE_BASIS_LOOKUP_CODE','BANK_CHARGE_DEDUCTION_TYPE','ALWAYS_TAKE_DISC_FLAG','EXCLUDE_FREIGHT_FROM_DISCOUNT','EXCLUDE_TAX_FROM_DISCOUNT','AUTO_CALCULATE_INTEREST_FLAG','VAT_CODE','VAT_REGISTRATION_NUM','PAYMENT_METHOD_LOOKUP_CODE','DELIVERY_CHANNEL_CODE','BANK_INSTRUCTION1_CODE','BANK_INSTRUCTION2_CODE','BANK_INSTRUCTION_DETAILS','SETTLEMENT_PRIORITY','PAYMENT_TEXT_MESSAGE1','PAYMENT_TEXT_MESSAGE2','PAYMENT_TEXT_MESSAGE3','IBY_BANK_CHARGE_BEARER','PAYMENT_REASON_CODE','PAYMENT_REASON_COMMENTS','REMIT_ADVICE_DELIVERY_METHOD','REMITTANCE_EMAIL','REMITTANCE_FAX','ATTRIBUTE_CATEGORY','ATTRIBUTE1','ATTRIBUTE2','ATTRIBUTE3','ATTRIBUTE4','ATTRIBUTE5','ATTRIBUTE6','ATTRIBUTE7','ATTRIBUTE8','ATTRIBUTE9','ATTRIBUTE10','ATTRIBUTE11','ATTRIBUTE12','ATTRIBUTE13','ATTRIBUTE14','ATTRIBUTE15','ATTRIBUTE16','ATTRIBUTE17','ATTRIBUTE18','ATTRIBUTE19','ATTRIBUTE20','ATTRIBUTE_DATE1','ATTRIBUTE_DATE2','ATTRIBUTE_DATE3','ATTRIBUTE_DATE4','ATTRIBUTE_DATE5','ATTRIBUTE_DATE6','ATTRIBUTE_DATE7','ATTRIBUTE_DATE8','ATTRIBUTE_DATE9','ATTRIBUTE_DATE10','ATTRIBUTE_TIMESTAMP1','ATTRIBUTE_TIMESTAMP2','ATTRIBUTE_TIMESTAMP3','ATTRIBUTE_TIMESTAMP4','ATTRIBUTE_TIMESTAMP5','ATTRIBUTE_TIMESTAMP6','ATTRIBUTE_TIMESTAMP7','ATTRIBUTE_TIMESTAMP8','ATTRIBUTE_TIMESTAMP9','ATTRIBUTE_TIMESTAMP10','ATTRIBUTE_NUMBER1','ATTRIBUTE_NUMBER2','ATTRIBUTE_NUMBER3','ATTRIBUTE_NUMBER4','ATTRIBUTE_NUMBER5','ATTRIBUTE_NUMBER6','ATTRIBUTE_NUMBER7','ATTRIBUTE_NUMBER8','ATTRIBUTE_NUMBER9','ATTRIBUTE_NUMBER10','GLOBAL_ATTRIBUTE_CATEGORY','GLOBAL_ATTRIBUTE1','GLOBAL_ATTRIBUTE2','GLOBAL_ATTRIBUTE3','GLOBAL_ATTRIBUTE4','GLOBAL_ATTRIBUTE5','GLOBAL_ATTRIBUTE6','GLOBAL_ATTRIBUTE7','GLOBAL_ATTRIBUTE8','GLOBAL_ATTRIBUTE9','GLOBAL_ATTRIBUTE10','GLOBAL_ATTRIBUTE11','GLOBAL_ATTRIBUTE12','GLOBAL_ATTRIBUTE13','GLOBAL_ATTRIBUTE14','GLOBAL_ATTRIBUTE15','GLOBAL_ATTRIBUTE16','GLOBAL_ATTRIBUTE17','GLOBAL_ATTRIBUTE18','GLOBAL_ATTRIBUTE19','GLOBAL_ATTRIBUTE20','GLOBAL_ATTRIBUTE_DATE1','GLOBAL_ATTRIBUTE_DATE2','GLOBAL_ATTRIBUTE_DATE3','GLOBAL_ATTRIBUTE_DATE4','GLOBAL_ATTRIBUTE_DATE5','GLOBAL_ATTRIBUTE_DATE6','GLOBAL_ATTRIBUTE_DATE7','GLOBAL_ATTRIBUTE_DATE8','GLOBAL_ATTRIBUTE_DATE9','GLOBAL_ATTRIBUTE_DATE10','GLOBAL_ATTRIBUTE_DATETIME1','GLOBAL_ATTRIBUTE_DATETIME2','GLOBAL_ATTRIBUTE_DATETIME3','GLOBAL_ATTRIBUTE_DATETIME4','GLOBAL_ATTRIBUTE_DATETIME5','GLOBAL_ATTRIBUTE_DATETIME6','GLOBAL_ATTRIBUTE_DATETIME7','GLOBAL_ATTRIBUTE_DATETIME8','GLOBAL_ATTRIBUTE_DATETIME9','GLOBAL_ATTRIBUTE_DATETIME10','GLOBAL_ATTRIBUTE_NUMBER1','GLOBAL_ATTRIBUTE_NUMBER2','GLOBAL_ATTRIBUTE_NUMBER3','GLOBAL_ATTRIBUTE_NUMBER4','GLOBAL_ATTRIBUTE_NUMBER5','GLOBAL_ATTRIBUTE_NUMBER6','GLOBAL_ATTRIBUTE_NUMBER7','GLOBAL_ATTRIBUTE_NUMBER8','GLOBAL_ATTRIBUTE_NUMBER9','GLOBAL_ATTRIBUTE_NUMBER10','PO_ACK_REQD_CODE','PO_ACK_REQD_DAYS','INVOICE_CHANNEL','PAYEE_SERVICE_LEVEL_CODE','EXCLUSIVE_PAYMENT_FLAG']
    textLine = chr(59).join([chr(39)+str(tmpColumn)+chr(39) for tmpColumn in tmpVector]) + '\n'
    textFile.write(textLine)
    linesWritten = 0
    for index, row in newReportDf.iterrows():
        tmpVector = row.to_list()
        textLine = chr(59).join([chr(39)+str(tmpColumn)+chr(39) for tmpColumn in tmpVector]) + '\n'
        textFile.write(textLine)
        linesWritten += 1
    textFile.close()
print(linesWritten, ' text lines written.')

### Step 11 - Validate the number of rows on file created

textFileToValidate = open(Path(workFolder,'3. SUPPLIER_SITES_INTERFACE_COPPEL(NEW).csv'), 'r', encoding='utf-8')
if textFileToValidate:
    numTextLines = 0
    while True:
        numTextLines += 1
        textLine = textFileToValidate.readline()
        print(textLine)
        if numTextLines >10: break
    inputTextFile.close()

validationDf = pd.read_csv(Path(workFolder,'3. SUPPLIER_SITES_INTERFACE_COPPEL(NEW).csv'), sep=';', quotechar=chr(39), encoding='utf-8')
validationDf.dtypes
print(len(validationDf.index), ' records found.')

### Step 12 - Additional Validation - duplicated names

sql_query = '''
SELECT DISTINCT NUMERO_PROVEEDOR,VENDOR_NAME,VENDOR_SITE_CODE FROM newReportDf
'''
SupplierVendorNameVendorSiteCodeDf = ps.sqldf(sql_query)
SupplierVendorNameVendorSiteCodeDf['NUMERO_PROVEEDOR'] = SupplierVendorNameVendorSiteCodeDf['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
SupplierVendorNameVendorSiteCodeDf['VENDOR_NAME'] = SupplierVendorNameVendorSiteCodeDf['VENDOR_NAME'].astype(pd.StringDtype())
SupplierVendorNameVendorSiteCodeDf['VENDOR_SITE_CODE'] = SupplierVendorNameVendorSiteCodeDf['VENDOR_SITE_CODE'].astype(pd.StringDtype())
print(len(SupplierVendorNameVendorSiteCodeDf.index), ' number of records.')
print(SupplierVendorNameVendorSiteCodeDf.head(10))

sql_query = '''
SELECT NUMERO_PROVEEDOR,VENDOR_SITE_CODE,COUNT(*) AS numObs 
FROM SupplierVendorNameVendorSiteCodeDf
GROUP BY NUMERO_PROVEEDOR,VENDOR_SITE_CODE
ORDER BY numObs DESC
'''
SupplierVendorSiteCodeDf = ps.sqldf(sql_query)
#
sql_query = '''
SELECT NUMERO_PROVEEDOR,VENDOR_SITE_CODE FROM SupplierVendorSiteCodeDf
WHERE numObs=1
'''
noDupsSupplierVendorSiteCodeDf = ps.sqldf(sql_query)
noDupsSupplierVendorSiteCodeDf['NUMERO_PROVEEDOR'] = noDupsSupplierVendorSiteCodeDf['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
noDupsSupplierVendorSiteCodeDf['VENDOR_SITE_CODE'] = noDupsSupplierVendorSiteCodeDf['VENDOR_SITE_CODE'].astype(pd.StringDtype())
print(len(noDupsSupplierVendorSiteCodeDf.index), ' number of non duplicate records.')
#
sql_query = '''
SELECT NUMERO_PROVEEDOR,VENDOR_SITE_CODE FROM SupplierVendorSiteCodeDf
WHERE numObs>1
'''
dupSupplierVendorSiteCodeDf = ps.sqldf(sql_query)
dupSupplierVendorSiteCodeDf['NUMERO_PROVEEDOR'] = dupSupplierVendorSiteCodeDf['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
dupSupplierVendorSiteCodeDf['VENDOR_SITE_CODE'] = dupSupplierVendorSiteCodeDf['VENDOR_SITE_CODE'].astype(pd.StringDtype())
print(len(dupSupplierVendorSiteCodeDf.index), ' number of duplicate records.')

dataSetWithUniqueRecordsStep1 = noDupsSupplierVendorSiteCodeDf.merge(SupplierVendorNameVendorSiteCodeDf, on=['NUMERO_PROVEEDOR','VENDOR_SITE_CODE'], how='left')
print(len(dataSetWithUniqueRecordsStep1.index), ' non duplicated supplier id, vendor site code and name')
print(dataSetWithUniqueRecordsStep1.head(10))
dataSetWithUniqueRecordsStep1.to_csv(Path(workFolder, 'uniqueSupplierVendorSitecodeandVendorNameStep1.csv'), encoding='utf-8')
dataSetToReview = dupSupplierVendorSiteCodeDf.merge(SupplierVendorNameVendorSiteCodeDf, on=['NUMERO_PROVEEDOR','VENDOR_SITE_CODE'], how='left')
print(len(dataSetToReview.index), ' duplicated supplier id, vendor site code and name')
print(dataSetToReview.head(10))
dataSetToReview.to_csv(Path(workFolder, 'duplicateSupplierVendorSitecodeandVendorName.csv'), encoding='utf-8')

sql_query = '''
SELECT NUMERO_PROVEEDOR,COUNT(*) AS numObs FROM dataSetWithUniqueRecordsStep1
GROUP BY NUMERO_PROVEEDOR
ORDER BY numObs DESC
'''
dataSetWithUniqueRecordsStep2 = ps.sqldf(sql_query)
dataSetWithUniqueRecordsStep2['NUMERO_PROVEEDOR'] = dataSetWithUniqueRecordsStep2['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
print(len(dataSetWithUniqueRecordsStep2.index), ' records')
#
sql_query1 = '''
SELECT NUMERO_PROVEEDOR FROM dataSetWithUniqueRecordsStep2
WHERE numObs=1
'''
uniqueSupplierIds = ps.sqldf(sql_query1)
uniqueSupplierIds['NUMERO_PROVEEDOR'] = uniqueSupplierIds['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
print(len(uniqueSupplierIds.index), ' unique records')
#
sql_query2 = '''
SELECT NUMERO_PROVEEDOR FROM dataSetWithUniqueRecordsStep2
WHERE numObs>1
'''
dupsSupplierIds = ps.sqldf(sql_query2)
dupsSupplierIds['NUMERO_PROVEEDOR'] = dupsSupplierIds['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
print(len(dupsSupplierIds.index), ' non unique records')

goldSupplierVscVendorName =
