==> 5_bankAccounts.ipynb <==
### Step 1. Load required packages

import pandas as pd
import pandasql as ps
from pathlib import Path

### Step 2 Load bank account data

sourceFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/5_Suppliers_Payees_BankAccounts'
filename = '5.2. SUPPLIER_BANK_ACCOUNTS_COPPEL.csv'
df = pd.read_csv(Path(sourceFolder,filename), sep=';', quotechar=chr(39), encoding='latin-1')

df['CLABE_NUMCHAR'] = 0
df['CLABE_VALIDATION'] = False 
df['ATTRIBUTE1'] = df['ATTRIBUTE1'].astype(str)
for index, row in df.iterrows():
    accountNum = row['ATTRIBUTE1'].replace('.0','')
    numChars = len(accountNum)
    df.loc[index, 'CLABE_NUMCHAR'] = numChars
    if numChars == 18:
        df.loc[index, 'CLABE_VALIDATION'] = True
print(df.head(10))

sql_query = '''
SELECT CLABE_VALIDATION,COUNT(*) AS NumCases FROM df
GROUP BY CLABE_VALIDATION
ORDER BY NumCases DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['CLABE_VALIDATION'], ', ', row['NumCases'])


sql_query = '''
SELECT CLABE_NUMCHAR,COUNT(*) AS NumCases FROM df
GROUP BY CLABE_NUMCHAR
ORDER BY CLABE_NUMCHAR DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['CLABE_NUMCHAR'],' :', row['NumCases'])

workFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/Work/5_Bank_Accounts'
ValidClabetDf = df.loc[df['CLABE_VALIDATION'] == True]
ValidClabetDf.to_csv(str(Path(workFolder,'ValidCabledDf.csv')), encoding='utf-8')

for column in list(ValidClabetDf.columns):
    print(column)

print(ValidClabetDf.head(10))

sql_query = '''
SELECT NUMERO_PROVEEDOR,ATTRIBUTE1,COUNT(*) AS numObs FROM ValidClabetDf
GROUP BY NUMERO_PROVEEDOR,ATTRIBUTE1
ORDER BY numObs DESC
'''
ValidClabeWithCaseNum = ps.sqldf(sql_query)
ValidClabeWithCaseNum['numCase'] = 0
numCases = 1
for index, row in ValidClabeWithCaseNum.iterrows():
    ValidClabeWithCaseNum.loc[index, 'numCase'] = numCases
    numCases = numCases+1
print(ValidClabeWithCaseNum.head(10))

ValidClabeWithCaseNum.to_csv(str(Path(workFolder,'ValidClabeWithCaseNum.csv')), encoding='utf-8')
