==> 5_2_bankAccounts.ipynb <==
##  Data from File 5.2 Bank Accounts

### Step 1. Load required packages

import pandas as pd
import pandasql as ps
from pathlib import Path

#### Step 2 - Get sample from source file

sourceFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/5_Suppliers_Payees_BankAccounts'
workFolder = '/content/drive/Shareddrives/Esperanto/Supports/Suppliers/Work/5_Bank_Accounts'
fileName = '5.2. SUPPLIER_BANK_ACCOUNTS_COPPEL.csv'
inputTextFile = open(Path(sourceFolder, fileName), 'r', encoding='latin-1')
if inputTextFile:
    numTextLines = 0
    while True:
        numTextLines += 1
        textLine = inputTextFile.readline()
        print(textLine)
        if numTextLines >10:
            break
    inputTextFile.close()

### Step 3 - Load bank account data

df = pd.read_csv(Path(sourceFolder,fileName), sep=';', quotechar=chr(39), encoding='latin-1')
df.dtypes

### Step 4 - Create a New Dataset (Modified) and Validate Bank Account

modDf = df
modDf['CLABE_VALIDATION'] = False
modDf['SISTEMA_LEGADO'] = modDf['SISTEMA_LEGADO'].astype(pd.StringDtype())
modDf['NUMERO_PROVEEDOR'] = modDf['NUMERO_PROVEEDOR'].astype(pd.StringDtype())
modDf['BANK_NAME'] = modDf['BANK_NAME'].astype(pd.StringDtype())
modDf['BRANCH_NAME'] = modDf['BRANCH_NAME'].astype(pd.StringDtype())
modDf['COUNTRY_CODE'] = modDf['COUNTRY_CODE'].astype(pd.StringDtype())
modDf['CURRENCY_CODE'] = modDf['CURRENCY_CODE'].astype(pd.StringDtype())
modDf['FOREING_PAYMENT_USE_FLAG'] = modDf['FOREING_PAYMENT_USE_FLAG'].astype(pd.StringDtype())
modDf['ATTRIBUTE1'] = modDf['ATTRIBUTE1'].astype(pd.StringDtype())
modDf['BANK_ACCOUNT_NUM'] = modDf['BANK_ACCOUNT_NUM'].astype(pd.StringDtype()) 
modDf['ATTRIBUTE1'] = modDf['ATTRIBUTE1'].astype(pd.StringDtype())
for index, row in modDf.iterrows():
    if len(str(row['ATTRIBUTE1']).replace('.0',''))==18:
        modDf.loc[index, 'CLABE_VALIDATION'] = True
#
modDf = modDf[modDf['NUMERO_PROVEEDOR'].notna()]
print('the orignal dataset contains ', len(df.index), ' records')
print('the modified dataset contains ', len(modDf.index), ' records.')
print(modDf.head(10))
print(modDf.dtypes)

### Step 5 - Creates a subset with elements whose taxt id and bank account are valid

subSetDf = modDf[modDf['CLABE_VALIDATION']==True]
for index, row in subSetDf.iterrows():
    bankAccountStr = str(row['BANK_ACCOUNT_NUM'])
    clabeAccountStr = str(row['ATTRIBUTE1'])
    subSetDf.at[index, 'BANK_ACCOUNT_NUM'] = bankAccountStr.replace('.0','')
    subSetDf.at[index, 'ATTRIBUTE1'] = clabeAccountStr.replace('.0','')
print(len(modDf.index) ,' records in original dataset')
print(len(subSetDf.index), ' records with a valid bank account (18 digits)')
print(subSetDf.head(10))

### Step 6 - Performs analysis for duplicate records based on taxt id or supplier id

sql_query = '''
SELECT NUMERO_PROVEEDOR,COUNT(*) AS numObs FROM subSetDf
GROUP BY NUMERO_PROVEEDOR
ORDER BY numObs DESC
'''
dupsAnalysis = ps.sqldf(sql_query)
print(len(dupsAnalysis.index), ' taxt ids in daset')
uniqueTaxdIds = dupsAnalysis[dupsAnalysis['numObs']==1]
print(len(uniqueTaxdIds.index), ' unique taxt ids')
duplicateTaxdIds = dupsAnalysis[dupsAnalysis['numObs']>1]
print(len(duplicateTaxdIds.index), ' duplicate taxt ids')
#
sql_query = '''
SELECT SISTEMA_LEGADO,COUNT(*) AS numObs FROM 
(SELECT SISTEMA_LEGADO,NUMERO_PROVEEDOR FROM subSetDf)
GROUP BY SISTEMA_LEGADO
ORDER BY numObs DESC
'''
distPerLegacy = ps.sqldf(sql_query)
for index, row in distPerLegacy.iterrows():
    print(row['SISTEMA_LEGADO'], ',', row['numObs'])

print('unique taxt ids: ')
print(uniqueTaxdIds.head(10))
print('duplicate taxt ids: ')
print(duplicateTaxdIds.head(100))

sql_query = '''
SELECT numObs,COUNT(*) AS NumCases FROM duplicateTaxdIds
GROUP BY numObs
ORDER BY numObs DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['numObs'], ' duplicates, ', row['NumCases'], ' cases')

### Step 7 - Perfoms analysis of distribution of accounts per supplier

sql_query = '''
SELECT DISTINCT NUMERO_PROVEEDOR,ATTRIBUTE1 FROM subSetDf
ORDER BY NUMERO_PROVEEDOR
'''
dataInScopeDf = ps.sqldf(sql_query)
print(len(dataInScopeDf.index), ' records in scope')
dataInScopeDf.to_csv(Path(workFolder, 'dataInScopeDfs.csv'), encoding='utf-8')
sql_query = '''
SELECT NUMERO_PROVEEDOR,COUNT(*) AS numObs FROM dataInScopeDf
GROUP BY NUMERO_PROVEEDOR
ORDER BY numObs DESC
'''
accountsPerSupplierDf = ps.sqldf(sql_query)
print(len(accountsPerSupplierDf.index), ' suppliers in dataset')
sql_query = '''
SELECT numObs,COUNT(*) AS numCases FROM accountsPerSupplierDf
GROUP BY numObs
ORDER BY numObs DESC
'''
result = ps.sqldf(sql_query)
for index, row in result.iterrows():
    print(row['numObs'], ', ', row['numCases'])
#
sql_query = '''
SELECT DISTINCT SISTEMA_LEGADO FROM subSetDf
''' 
legacySystemsInScopeDf = ps.sqldf(sql_query)
print(len(legacySystemsInScopeDf.index), ' legacy sistems in scope')
legacySystemsInScopeDf.to_csv(Path(workFolder, 'legacySystemsInScopeDfs.csv'), encoding='utf-8')

### Step 8 - Create new reporting structure from subset

newReportDf = subSetDf[['SISTEMA_LEGADO','NUMERO_PROVEEDOR','FEEDER_IMPORT_BATCH_ID','TEMP_EXT_PAYEE_ID','TEMP_EXT_BANK_ACCT_ID',
'BANK_NAME','BRANCH_NAME','COUNTRY_CODE','BANK_ACCOUNT_NAME','BANK_ACCOUNT_NUM','CURRENCY_CODE','FOREING_PAYMENT_USE_FLAG',
'START_DATE','END_DATE','IBAN','CHECK_DIGITS','BANK_ACCOUNT_NAME_ALT','BANK_ACCOUNT_TYPE','ACCOUNT_SUFFIX','DESCRIPTION',
'AGENCY_LOCATION_CODE','EXCHANGE_RATE_AGREEMENT_NUM','EXCHANGE_RATE_AGREEMENT_TYPE','EXCHANGE_RATE','SECONDARY_ACCOUNT_REFERENCE',
'ATTRIBUTE_CATEGORY','ATTRIBUTE1','ATTRIBUTE2','ATTRIBUTE3','ATTRIBUTE4','ATTRIBUTE5','ATTRIBUTE6','ATTRIBUTE7','ATTRIBUTE8',
'ATTRIBUTE9','ATTRIBUTE10','ATTRIBUTE11','ATTRIBUTE12','ATTRIBUTE13','ATTRIBUTE14','ATTRIBUTE15']]
print(len(newReportDf.index), ' in new report structure')

### Step 9 - Create new textfile

print(newReportDf.dtypes)

newReportDf.fillna('', inplace=True)
with open(Path(workFolder,'5.2 SUPPLIER_BANK_ACCOUNTS_COPPEL(NEW).csv'), 'w', encoding='utf-8') as textFile:
    tmpVector = ['SISTEMA_LEGADO','NUMERO_PROVEEDOR','FEEDER_IMPORT_BATCH_ID','TEMP_EXT_PAYEE_ID','TEMP_EXT_BANK_ACCT_ID',
                 'BANK_NAME','BRANCH_NAME','COUNTRY_CODE','BANK_ACCOUNT_NAME','BANK_ACCOUNT_NUM','CURRENCY_CODE','FOREING_PAYMENT_USE_FLAG',
                 'START_DATE','END_DATE','IBAN','CHECK_DIGITS','BANK_ACCOUNT_NAME_ALT','BANK_ACCOUNT_TYPE','ACCOUNT_SUFFIX','DESCRIPTION',
                 'AGENCY_LOCATION_CODE','EXCHANGE_RATE_AGREEMENT_NUM','EXCHANGE_RATE_AGREEMENT_TYPE','EXCHANGE_RATE','SECONDARY_ACCOUNT_REFERENCE',
                 'ATTRIBUTE_CATEGORY','ATTRIBUTE1','ATTRIBUTE2','ATTRIBUTE3','ATTRIBUTE4','ATTRIBUTE5','ATTRIBUTE6','ATTRIBUTE7','ATTRIBUTE8',
                 'ATTRIBUTE9','ATTRIBUTE10','ATTRIBUTE11','ATTRIBUTE12','ATTRIBUTE13','ATTRIBUTE14','ATTRIBUTE15']
    textLine = chr(59).join([chr(39)+str(tmpColumn)+chr(39) for tmpColumn in tmpVector]) + '\n'
    textFile.write(textLine)
    linesWritten = 0
    for index, row in newReportDf.iterrows():
        tmpVector = row.to_list()
        textLine = chr(59).join([chr(39)+str(tmpColumn)+chr(39) for tmpColumn in tmpVector]) + '\n'
        textFile.write(textLine)
        linesWritten += 1
    textFile.close()
print(linesWritten, ' text lines written.')

### Step 10 - Validate the number of rows on file created

textFileToValidate = open(Path(workFolder,'5.2 SUPPLIER_BANK_ACCOUNTS_COPPEL(NEW).csv'), 'r', encoding='utf-8')
if textFileToValidate:
    numTextLines = 0
    while True:
        numTextLines += 1
        textLine = textFileToValidate.readline()
        print(textLine)
        if numTextLines >10: break
    inputTextFile.close()

validationDf = pd.read_csv(Path(workFolder,'5.2 SUPPLIER_BANK_ACCOUNTS_COPPEL(NEW).csv'), sep=';', quotechar=chr(39), encoding='utf-8')
validationDf.dtypes
print(len(validationDf.index), ' records found.')
